data_analytics/requirements.txt
pandas>=2.0.0
pymongo>=4.5.0
python-dotenv>=1.0.0
bson>=0.5.10
numpy>=1.24.0
fpdf2>=2.7.0